Rename event-service file to event-bus
