+++ server/src/core/event/index.ts
@@ -1,1 +1,1 @@
-export { EventBus, createEvent } from './event-service';
+export { EventBus, createEvent } from './event-bus';
